third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/ui-source-code-metadata-expected.txt

Verify that network UISourceCode has metadata.


Running: testCachedResource
Metadata for UISourceCode: http://127.0.0.1:8000/devtools/sources/resources/style.css
    content size: 63
    modification time: <Date>

Running: testDynamicResource
Metadata for UISourceCode: http://127.0.0.1:8000/devtools/sources/resources/script.js
    content size: 31
    modification time: <Date>

Running: testInlinedSourceMapSource
Metadata for UISourceCode: http://127.0.0.1:8000/devtools/sources/resources/style.scss
    content size: 26
    modification time: null
